backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-multipart
pillow
tensorflow
python-dotenv